style: add new style to text-editor
