--- src/components/quill-editor/formats.ts
@@ -1,20 +1,0 @@
-export const formats = [
-  'bold',
-  'color',
-  'font',
-  'code',
-  'italic',
-  'link',
-  'size',
-  'strike',
-  'script',
-  'underline',
-  'blockquote',
-  'header',
-  'indent',
-  'list',
-  'align',
-  'direction',
-  'code-block',
-  'formula',
-]
